% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b1}
\alias{b1}
\title{3d topological surface measurements for one land of bullet}
\format{An object of class \code{list} of length 2.}
\usage{
b1
}
\description{
Some more info - not sure at the moment which bullet this is. Describe structure.
}
\keyword{datasets}

