% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bulletAlign}
\alias{bulletAlign}
\title{Align two surface cross cuts according to maximal correlation}
\usage{
bulletAlign(data, value = "l30")
}
\arguments{
\item{data}{data frame consisting of at least two surface crosscuts as given by function \code{bulletSmooth}.}

\item{value}{string of the variable to match. Defaults to l30, the variable returned from function \code{bulletSmooth}.}
}
\value{
same data frame as input, but y vectors are aligned for maximal correlation between the
}
\description{
Align two surface cross cuts according to maximal correlation
}

