% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bulletCheckCrossCutOld}
\alias{bulletCheckCrossCutOld}
\title{keep for backup right now}
\usage{
bulletCheckCrossCutOld(path, distance = 25, x = seq(100, 225, by =
  distance))
}
\description{
keep for backup right now
}

