% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bulletGetMaxCMS}
\alias{bulletGetMaxCMS}
\title{Identify the number of maximum CMS between two bullet lands}
\usage{
bulletGetMaxCMS(bullet1, bullet2, crosscut = 100, thresholds = seq(0.3, 1.5,
  by = 0.05))
}
\arguments{
\item{bullet1, }{bullet2 paths to two lands}

\item{crosscut}{integer value specifying which surface crosscut to use for the match}

\item{thresholds}{vector of potential thresholds to choose from for optimizing the number of CMS}
}
\value{
list of matching parameters, data set of the identified striae, and the aligned data sets.
}
\description{
Identify the number of maximum CMS between two bullet lands
}

