% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bulletPickThreshold}
\alias{bulletPickThreshold}
\title{Pick threshold based on maximum number of CMS}
\usage{
bulletPickThreshold(data, thresholds)
}
\arguments{
\item{data}{data frame consisting of at least two surface crosscuts as given by function \code{bulletAlign}.}

\item{thresholds}{vector of thresholds under consideration. Typically in the range between 0.3 and 1.5 (with 5 being the maximum).}
}
\value{
threshold value that maximizes CMS
}
\description{
Pick threshold based on maximum number of CMS
}

