% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CMS}
\alias{CMS}
\title{Table of the number of consecutive matches}
\usage{
CMS(match)
}
\arguments{
\item{match}{is a Boolean vector of matches/non-matches}
}
\value{
a table of the number of the CMS and their frequencies
}
\description{
Table of the number of consecutive matches
}

