% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{fortify}
\alias{fortify}
\title{Convert a list of x3d file into a data frame}
\usage{
fortify(x3d)
}
\arguments{
\item{x3d}{a file in x3d format as return by function read.x3d}
}
\value{
data frame with variables x, y, and value
}
\description{
x3d format consists of a list with header info and a 2d matrix of scan depths.
Fortify turn the matrix into a variable within a data frame, using the parameters of the header as necessary.
}

