% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fortify_x3p}
\alias{fortify_x3p}
\title{Convert a list of x3d file into a data frame}
\usage{
fortify_x3p(x3d)
}
\arguments{
\item{x3d}{a file in x3d format as return by function read.x3d}
}
\value{
data frame with variables x, y, and value
}
\description{
x3d format consists of a list with header info and a 2d matrix of scan depths. 
fortify_x3p turn the matrix into a variable within a data frame, using the parameters of the header as necessary.
}

