% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_crosscut}
\alias{get_crosscut}
\title{Read a crosscut from a 3d surface file}
\usage{
get_crosscut(path, x = 243.75)
}
\arguments{
\item{path}{path to an x3p file}

\item{x}{level of the crosscut to be taken. If this level does not exist, the crosscut with the closest level is returned.}
}
\value{
data frame
}
\description{
Read a crosscut from a 3d surface file
}

