% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getTwist}
\alias{getTwist}
\title{Estimate the twist in a bullet land}
\usage{
getTwist(path, bullet = NULL, transpose = FALSE, twistlimit = NULL,
  cutoff = 0.75)
}
\arguments{
\item{path}{to a file in x3p format}

\item{bullet}{data in x3p format as returned by function read.x3p}

\item{transpose}{If TRUE, transpose the matrix}
}
\value{
numeric value estimating the twist
}
\description{
Estimation of the twist in a barrel follows roughly the process described by Chu et al (2010).
At the moment, twist is estimated from a single land - but the twist should be the same for the whole barrel. Therefore all lands of the same barrel should
have the same twist.
A note on timing: at the moment calculating the twist rate for a bullet land takes several minutes.
}

