% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read.x3pplus}
\alias{read.x3pplus}
\title{Read X3P File}
\usage{
read.x3pplus(fpath, transpose = FALSE)
}
\arguments{
\item{fpath}{The file path to the x3p file}

\item{transpose}{If TRUE, transpose the resulting matrix}
}
\description{
Read X3P File
}

