% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{import}
\name{reexports}
\alias{read.x3p}
\alias{reexports}
\alias{xmlRoot}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{x3pr}{\code{\link[x3pr]{read.x3p}}}

  \item{XML}{\code{\link[XML]{xmlRoot}}}
}}
\keyword{internal}

