% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x3p.R
\name{sample.x3p}
\alias{sample.x3p}
\title{Sample every X element of a data frame}
\usage{
sample.x3p(dframe, byxy = c(2, 2))
}
\arguments{
\item{dframe}{data frame with x and y variable}

\item{byxy}{(vector) of numeric value indicating the sapling resolution. If a single number, the same resolution is used for x and y.}
}
\value{
subset of the input variable
}
\description{
Sample every X element of a data frame in x and y direction
}

