% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{striation_identify}
\alias{striation_identify}
\title{Identify striation marks across two bullets}
\usage{
striation_identify(data, threshold = 0.75)
}
\arguments{
\item{data}{dataset containing crosscuts of (exactly?) two bullets as given by \code{processBullets}.}

\item{threshold}{where should the smoothed values be cut? Typically, residuals from the smooth have values in (-5,5). A default value of 0.75 is taken.}
}
\value{
a data frame with information on all of the identified striation marks, and whether they match across the two bullets.
}
\description{
Identify striation marks across two bullets
}

