% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{striation_identifyXXX}
\alias{striation_identifyXXX}
\title{Match striation marks across two cross sections based on previously identified peaks and valleys}
\usage{
striation_identifyXXX(lines1, lines2)
}
\arguments{
\item{lines1, }{lines2 data frames as returned from get_peaks function. data frames are expected to have 
the following variables: xmin, xmax, group, type, bullet, heights}
}
\description{
Match striation marks across two cross sections based on previously identified peaks and valleys
}

