% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bulletCheckCrossCut}
\alias{bulletCheckCrossCut}
\title{Identifying a reliable cross section}
\usage{
bulletCheckCrossCut(path, distance = 25, xlimits = c(50, 500),
  minccf = 0.9)
}
\arguments{
\item{path}{path to an x3p file}

\item{distance}{positive numeric value indicating the distance between cross sections to use for a comparison}

\item{xlimits}{vector of values between which to check for cross sections in a stable region}

\item{minccf}{minimal value of cross correlation to indicate a stable region}
}
\description{
Should be changed: x should just indicate lower and upper limit. That is cleaner and should speed things up as well.
}

