% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bulletGetMaxCMSXXX}
\alias{bulletGetMaxCMSXXX}
\title{Identify the number of maximum CMS between two bullet lands}
\usage{
bulletGetMaxCMSXXX(bullet1, bullet2, crosscut1, crosscut2, span = 35)
}
\arguments{
\item{bullet1}{path to x3p file of the first bullet-land.}

\item{bullet2}{path to x3p file of the second bullet-land.}

\item{crosscut1}{cross section at which to evaluate the match for the first bullet-land surface.}

\item{crosscut2}{cross section at which to evaluate the match for the second bullet-land surface}

\item{span}{positive number  for the smoothfactor to use for assessing peaks.}
}
\value{
list of matching parameters, data set of the identified striae, and the aligned data sets.
}
\description{
Identify the number of maximum CMS between two bullet lands
}

