% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bulletGetMaxCMSXXX}
\alias{bulletGetMaxCMSXXX}
\title{Identify the number of maximum CMS between two bullet lands}
\usage{
bulletGetMaxCMSXXX(bullet1, bullet2)
}
\arguments{
\item{bullet1, }{bullet2 load from bullets.RData}
}
\value{
list of matching parameters, data set of the identified striae, and the aligned data sets.
}
\description{
Identify the number of maximum CMS between two bullet lands
}

