% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_grooves}
\alias{get_grooves}
\title{Find the grooves of a bullet land}
\usage{
get_grooves(bullet, smoothfactor = 35, smoothplot = FALSE, adjust = 10)
}
\arguments{
\item{adjust}{}
}
\description{
Find the grooves of a bullet land
}

