% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{processBullets}
\alias{processBullets}
\title{Process x3p file}
\usage{
processBullets(paths, x = 100, check = FALSE)
}
\arguments{
\item{paths}{file paths to the x3p files}

\item{x}{(vector) of surface crosscuts to process.}

\item{check}{boolean indicating whether the crosscut should be checked for stability. Set to FALSE by default.}
}
\value{
data frame
}
\description{
x3p file of a 3d topological bullet surface is processed at surface crosscut x, 
the bullet grooves in the crosscuts are identified and removed, and a loess smooth 
is used (see \code{?loess} for details) to remove the big structure.
}

